\name{elisa.R.gh}
\alias{elisa.mean.distr.gh}
\alias{elisa.R.gh}
\docType{data}
\title{
ELISA prior for gh Parameterization
}
\description{
Priors.
}
\format{
elisa.R.gh is a 5x5 diagonal matrix. rowNames/colNames: "c" "d" "g" "logh" "logf"
elisa.mean.distr.gh is a 2x5 matrix. colNames: "c" "d" "g" "logh" "logf". rowNames: mean, prec
}
