\name{hier.model.ex.2}
\alias{hier.model.ex.2}
\docType{data}
\title{
An example for hierarchical modeling used in Fong, Yu et al. 
}
\description{
An example for hierarchical modeling used in Fong, Yu et al. 
}
\format{
  A data frame with 120 observations on the following 8 variables.
  \describe{
    \item{\code{assay_id}}{a character vector. Assay identifier.}
    \item{\code{analyte}}{a character vector. Substance to be measured.}
    \item{\code{expected_conc}}{a numeric vector. The expected concentrations of a sample.}
    \item{\code{dilution}}{a numeric vector. The dilution factor of a sample.}
    \item{\code{fi}}{a numeric vector. Fluoresence intensity readout.}
    \item{\code{well_role}}{a character vector. Defines the role of a well. See ?ncal for more information.}
    \item{\code{sample_id}}{a character vector. Sample identifier.}
    \item{\code{replicate}}{a numeric vector. Techinical replicate identifier.}
  }
}
\references{
Fong, Y., Yu, X., Sebestyen, K., Gilbert, P. and Self, S. (2013) nCal: a R package for nonlinear calibration. Submitted.
}
