\name{p.eotaxin}
\alias{p.eotaxin}
\docType{data}
\title{
5PL parameters from an Eotaxin dataset.
}
\description{
5PL parameters from an Eotaxin dataset.
}
\format{
  The format is:
 num [1:6, 1:12] 4.37 4.3 4.58 4.74 4.31 ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:6] "1" "2" "3" "4" ...
  ..$ : chr [1:12] "c" "d" "loge" "logmb" ...
}
