\name{dat.QIL3}
\alias{dat.QIL3}
\docType{data}
\title{
An example for hierarchical modeling used in Fong, Wakefield, De Rosa, Frahm (2012)
}
\description{
An example for hierarchical modeling used in Fong, Wakefield, De Rosa, Frahm (2012)
}
\format{
  A data frame with 120 observations on the following 13 variables.
  \describe{
    \item{\code{well}}{a character vector}
    \item{\code{assay_id}}{a character vector}
    \item{\code{analyte}}{a character vector}
    \item{\code{well_role}}{a character vector}
    \item{\code{beadct}}{a numeric vector}
    \item{\code{ptid}}{a numeric vector}
    \item{\code{visit}}{a numeric vector}
    \item{\code{thawdt}}{a character vector}
    \item{\code{dilution}}{a numeric vector}
    \item{\code{expected_conc}}{a numeric vector}
    \item{\code{fi}}{a numeric vector}
    \item{\code{sample_id}}{a logical vector}
    \item{\code{replicate}}{a numeric vector}
  }
}
\references{
Fong, Y., Wakefield, J., DeRosa, S., Frahm, N. (2012) A robust Bayesian random effects model for nonlinear calibration problems, Biometrics, 68:1103-1112.
}
