\name{hier.model.ex.2}
\alias{hier.model.ex.2}
\docType{data}
\title{
An example for hierarchical modeling used in Fong, Yu et al. 
}
\description{
An example for hierarchical modeling used in Fong, Yu et al. 
}
\format{
  A data frame with 120 observations on the following 8 variables.
  \describe{
    \item{\code{assay_id}}{a character vector}
    \item{\code{analyte}}{a character vector}
    \item{\code{expected_conc}}{a numeric vector}
    \item{\code{dilution}}{a numeric vector}
    \item{\code{fi}}{a numeric vector}
    \item{\code{well_role}}{a character vector}
    \item{\code{sample_id}}{a character vector}
    \item{\code{replicate}}{a numeric vector}
  }
}
\references{
Fong, Y., Yu, X., Sebestyen, K., Gilbert, P. and Self, S. (2013) nCal: a R package for nonlinear calibration. Submitted.
}
